+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+### Changed
+- do not re-assign r_th and c_th values when loading/creating a transistor object (do not run calc_thermal_params())
+
 ## [0.3.2] - 2021-02-02
 ### Added
 - Add Rohm to manufacturer list